+++ trimesh/exchange/assimp.py
