+++ trimesh/exchange/gltf.py
