+++ trimesh/exchange/load.py
