+++ trimesh/exchange/misc.py
