+++ trimesh/exchange/openctm.py
