+++ trimesh/exchange/ply.py
