+++ trimesh/exchange/stl.py
