+++ trimesh/exchange/threemf.py
