+++ trimesh/exchange/wavefront.py
