+++ trimesh/exchange/xml_based.py
